[Feat] 메인 기기대응 (#22)
