+++ WAL/WAL/Network/Service/Main/MainService.swift
@@ -49,7 +49,7 @@ extension MainService: TargetType {
     
     var headers: [String : String]? {
         return ["Content-Type": "application/json",
-                "accesstoken": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6MSwibmlja25hbWUiOiLsmYjrv6EiLCJlbWFpbCI6ImNob2N3MDQwMkBnbWFpbC5jb20iLCJzb2NpYWwiOiJrYWthbyIsImlhdCI6MTY1OTA5MjYyNiwiZXhwIjoxNjYwMzAyMjI2LCJpc3MiOiJjaGFud29vIn0.MOmYnBunDTrKXOl7eiI7AICcl8sgrS63m6wCH_dA8Vc"]
+                "accesstoken": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6OCwibmlja25hbWUiOiLro6jtnawiLCJlbWFpbCI6Imp3eWFuZGtyZ0BrYWthby5jb20iLCJzb2NpYWwiOiJrYWthbyIsImlhdCI6MTY2MDM5MzU1MCwiZXhwIjoxNjYxNjAzMTUwLCJpc3MiOiJjaGFud29vIn0.GetbS7CzKeUaCna4IIi_e2F4p8ZfLukES-N83RoxKKE"]
     }
 }
